add SkillSelect component
